cmd/compile: load results into registers on open defer return path
When a function panics then recovers, it needs to return to the
caller with named results having the correct values. For
in-register results, we need to load them into registers at the
defer return path. We already generate correct code for non-open-
coded defers. This CL generates instructions that load the result
registers in open defer return path.

Updates #40724.

Change-Id: I1f60514da644fd5fb4b4871a1153c62f42927282