cmd/link: enforce larger alignment on split ppc64 text sections
When laying out, we lazily increase the alignment for text sections
as symbols are found requiring more. This works if the start of
the virtual address (VA) is aligned sufficiently. This is a problem
when we split sections, and the start of the new text section is
not sufficiently aligned. The section start will be rounded up,
and some text data is lost if we encounter a later symbol with higher
alignment requirements.

This alignment adjust happens later in ld.(*Link).address which rounds
the VA to match alignment. If the VA wasn't already aligned, this
sometimes causes a sanity check failure against the number of
relocations streamed out for the previous text segment.

For now, align the start of any newly split text section to 64B.
This should be the highest alignment encountered when linking
ppc64 text segments.  Similarly, the xcoff size calculations need
to be improved to handle padding bytes when function symbols have
been aligned.  This is done by tracking the largest valid VA offset
encountered when placing symbols.

Change-Id: Iefef09a1ee7c963fb8dfce2288a084a95cb77fca