cmd/compile: pre-spill pointers in aggregate-typed register args
There's a problem in liveness, where liveness of any part of an aggregate
keeps the whole aggregate alive, but the not-live parts don't get spilled.
A better fix is to change liveness to work pointer-by-pointer, but that is
also a riskier, trickier fix.

To avoid this, in the case of

(1) an aggregate input parameter
(2) containing pointers
(3) passed in a registers

pre-spill the pointers.

Updates #40724.

Change-Id: I6beb8e0a353b1ae3c68c16072f56698061922c04