bytes: document Buffer.Bytes semantics more clearly
The Bytes function returns an alias into the internal buffer,
which allows users to do certain advanced operations assuming the
semantics of Bytes is sufficiently well defined.

Some of the advanced operations (e.g., mutating the unread buffer)
to affect future reads is already documented, but the use of Bytes
to obtain a buffer for writing is not documented. So document it.

Fixes #42986

Change-Id: I0109b5a51123552bc85ad53a1c17fba169a018b4