runtime: hide externalthreadhandler SP changes from nosplit check
externalthreadhandler is immediately called on a new OS thread for
console controls and profiling, with an OS allocated stack (specified as
1MB on 32bit, 2MB on 64bit by the linker).

This function creates an M and G on the stack, which together exceed the
entire nosplit stack limit on 64-bit, and are extremely close to the
limit on 32-bit.

This is nosplit since it is an unsafe non-Go context, but there is
plenty of stack space, so the nosplit stack depth check is not relevant,
so we want to skip the check.

On arm64, we already play assembly games to trick the check. Expand the
comments there.

On 386 and amd64, we perform standard adjustments, but the nosplit check
doesn't track them regardless. Expand the comments there.

On arm, we perform standard adjustments, and the nosplit check does
track them. Adding as little as 4-bytes to either M or G will trigger a
nosplit overflow. Play the same games as arm64 so that the next M or G
change doesn't break this.

Change-Id: I7796a850a4a11dabc14dd2812b793de389e82c0e