cmd/compile: provide alignment for globals to linker
Currently, the assembler uses heuristics to guess the alignment for
various linker symbols. This is okay for typical variables, but it
fails for variables that require greater-than-natural alignment (i.e.,
16-byte alignment or more).

This CL changes the compiler to explicitly provide alignment
information for global variables. Eventually, the compiler should
provide explicit alignments for everything.

Change-Id: I74795f3c9618e4dfa6ac92f25ad4a728789c3661