cmd/compile: move to new unstable export version, keep reading previous version
I added constants for the previous export versions, and for the final
generics export version. I also added a range for unstable versions,
which we can use as we are adjusting the export format for generics. Put
in the needed conditionals to be able to read old versions.

DO NOT SUBMIT

Change-Id: I166d17d943e07951aa752562e952b067704aeeca