testing: add -shuffle flag to alter execution order of tests and benchmarks
This CL adds a new flag to the testing package and the go test command
which enables randomizing the order of execution for tests and
benchmarks. This can be useful for identifying unwanted dependencies
between test and benchmark functions.
The flag is off by default. If `-shuffle` is set to `on` then the system
clock will be used as the seed value. If `-shuffle` is set to an integer
n, then n will be used as the seed value. In both cases, the seed will
be reported before before any runs so that failures can be reproduced
later on.

Fixes #28592

Change-Id: I62e7dfae5f63f97a0cbd7830ea844d9f7beac335