runtime: crash the GC when clobberdead pointer is seen
When -clobberdead compiler flag is set, the compiler inserts
instructions that set dead slots a specific value. If the GC sees
this value as a live pointer, something is probably wrong. Crash.

DO NOT SUBMIT. Need some way to make sure it is not in some valid
non-Go address space (e.g. C allocated, or user mmap).

Change-Id: Ica687b132b5d3ba2a062500d13264fa730405d11