cmd/compile: export the bound type params, re-import in types1/types2
Add the methods of the bound to the export format for a typeparam. We
then import the bound for types1 and types2, so types2 can do the proper
bound (constraint) check on the type args. We are not yet exporting type
lists, since those are changing and are not yet exposed via the types2
API. This change is really just a prototype, since we really need to
export a whole interface for the typeparam bound, including embedded
interfaces and union elements (when they are exported), etc.

Added new test stringimp.go

Allow OXDOT to be exported by the new importer. For generics, we will
have to be able to export pre-transformed and post-transformed ops,
since generic functions often cannot be transformed until the specific
type is known.  (This specific op is needed for the stringimp.go case.)

Change-Id: Ie5ddd350266292e269b89c67d0a581d6938beef3