cmd/dist: use goTest() to construct commands for more tests
Currently, most dist tests use (*tester).goTest() to construct "go
test" command line arguments, but a few don't, most notably the std
tests. The most noticeable effect of this is that we don't pass
-count=1 to these tests, which means running run.bash repeatedly will
just use the cached test results for the std tests and a few others.
Indeed, goTest() was originally introduced in CL 80735 with the intent
to disable test caching in run.bash.

This CL uses goTest() everywhere to construct the "go test" command line.

Change-Id: I96ece34dbeddfb142738c679a8b57aaa00bda6b0